requirements.txt
alabaster==0.7.12
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
docutils==0.17.1
idna==3.3
imagesize==1.2.0
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.3
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
packaging==21.2
Pillow==8.4.0
Pygments==2.10.0
pyparsing==2.4.7
PySide6==6.2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
shiboken6==6.2.1
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.7
